Removes Bibliography.
Like most of the world and unlike academia, we are allowed to live in the present and not in the past so we can just use hyperlinks (a technology from the 80s) directly rather than having to author a bibliography as a separate task.

In general, anytime you want to link out to somewhere it is best to just turn some text into a link and move on.  No need to write a bibliography, as no one but professors in school actually cares and they are less useful and harder to use than links.